tooling/aprl-automate/requirements.txt
azure-mgmt-resourcegraph
azure-mgmt-resource
azure-cli-core
azure-identity

pandas
argparse
openpyxl
